refactor(services): add browser service, update minimum supported browser versions, remove ungeneric services
