+++ packages/blockchain-wallet-v4-frontend/src/modals/SimpleBuy/EnterAmount/Checkout/template.success.tsx
@@ -19,13 +19,12 @@ import { Form } from 'components/Form'
 import { model } from 'data'
 import { convertStandardToBase } from 'data/components/exchange/services'
 import { SBCheckoutFormValuesType, SwapBaseCounterTypes } from 'data/types'
-import ErrorCodeMappings from 'services/ErrorCodeMappings'
 import { CRYPTO_DECIMALS, FIAT_DECIMALS, formatTextAmount } from 'services/forms'
 
 import Scheduler from '../../../RecurringBuys/Scheduler'
 import { Row } from '../../../Swap/EnterAmount/Checkout'
 import CryptoItem from '../../CryptoSelection/CryptoSelector/CryptoItem'
-import { BuyOrSell } from '../../model'
+import { BuyOrSell, ErrorCodeMappings } from '../../model'
 import Failure from '../template.failure'
 import { Props as OwnProps, SuccessStateType } from '.'
 import ActionButton from './ActionButton'
