+++ packages/blockchain-wallet-v4-frontend/src/modals/Swap/EnterAmount/index.tsx
@@ -10,10 +10,10 @@ import { FlyoutWrapper } from 'components/Flyout'
 import { selectors } from 'data'
 import { RootState } from 'data/rootReducer'
 import { InitSwapFormValuesType, SwapAccountType, SwapCoinType } from 'data/types'
-import checkAccountZeroBalance from 'services/CheckAccountZeroBalance'
 
 import { Props as BaseProps, SuccessStateType as SuccessType } from '..'
 import { BalanceRow, Border, Option, OptionTitle, OptionValue, TopText } from '../components'
+import { checkAccountZeroBalance } from '../model'
 import Checkout from './Checkout'
 import getData from './selectors'
 import Failure from './template.failure'
