+++ packages/blockchain-wallet-v4-frontend/src/scenes/Settings/Preferences/CryptoLinkHandling/index.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import { FormattedMessage } from 'react-intl'
 import { connect } from 'react-redux'
-import Bowser from 'bowser'
 import { bindActionCreators } from 'redux'
 import styled from 'styled-components'
 
@@ -14,11 +13,9 @@ import {
   SettingSummary
 } from 'components/Setting'
 import { actions, model } from 'data'
+import { isBrowserSafari } from 'services/browser'
 
-const browser = Bowser.getParser(window.navigator.userAgent)
-const isSafari = browser.satisfies({
-  safari: '>8'
-})
+const isSafari = isBrowserSafari()
 
 const TextWrapper = styled(Text)`
   a {
